// apps/backend/src/core/enums/task-priority.enum.ts
export enum TaskPriority {
  URGENT_IMPORTANT = 'urgent_important', // Do ti (Green)
  NOT_URGENT_IMPORTANT = 'not_urgent_important', // Schedule it (Orange)
  URGENT_NOT_IMPORTANT = 'urgent_not_important', // Delegate it (Yellow)
  NOT_URGENT_NOT_IMPORTANT = 'not_urgent_not_important', // Delete it (Red)
}